Extract some common code
